requirements.txt
Flask
Flask-Login
Flask-WTF
SQLAlchemy
Werkzeug
openai
gunicorn
python-dotenv
flask_sqlalchemy
requests
pillow
